# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI1Net API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  clerkId: string;
  email: string;
  name?: string | null;
  avatar?: string | null;
  tokenBalance: number;
  createdAt: string;
}

export interface UpdateUserBody {
  name?: string;
  avatar?: string;
}

export type AICategory = (typeof AICategory)[keyof typeof AICategory];

export const AICategory = {
  TEXT: "TEXT",
  IMAGE: "IMAGE",
  VIDEO: "VIDEO",
  CODE: "CODE",
  VOICE: "VOICE",
  MULTIMODAL: "MULTIMODAL",
} as const;

export interface AIProvider {
  id: string;
  name: string;
  description?: string | null;
  website?: string | null;
  toolCount: number;
}

export interface AITool {
  id: string;
  name: string;
  slug: string;
  description: string;
  category: AICategory;
  providerId: string;
  providerName: string;
  pricePerUse: number;
  createdAt: string;
}

export interface AIRequestBody {
  toolSlug: string;
  input: string;
  model?: string | null;
}

export type UsageStatus = (typeof UsageStatus)[keyof typeof UsageStatus];

export const UsageStatus = {
  SUCCESS: "SUCCESS",
  FAILED: "FAILED",
  PENDING: "PENDING",
} as const;

export interface AIRequestResponse {
  usageId: string;
  output: string;
  tokensUsed: number;
  remainingBalance: number;
  status: UsageStatus;
}

export interface AIUsage {
  id: string;
  toolId: string;
  toolName: string;
  toolSlug: string;
  category: AICategory;
  input: string;
  output?: string | null;
  tokensUsed: number;
  status: UsageStatus;
  createdAt: string;
}

export interface PaginatedUsage {
  data: AIUsage[];
  total: number;
  limit: number;
  offset: number;
}

export type UsageStatsTopCategoriesItem = {
  category: AICategory;
  count: number;
};

export interface UsageStats {
  totalRequests: number;
  totalTokensSpent: number;
  requestsToday: number;
  tokensSpentToday: number;
  successRate: number;
  topCategories: UsageStatsTopCategoriesItem[];
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  SPEND: "SPEND",
  EARN: "EARN",
  STAKE: "STAKE",
  UNSTAKE: "UNSTAKE",
  REWARD: "REWARD",
} as const;

export interface TokenBalance {
  balance: number;
  stakedAmount: number;
  availableBalance: number;
}

export interface TokenTransaction {
  id: string;
  type: TransactionType;
  amount: number;
  metadata?: string | null;
  createdAt: string;
}

export interface PaginatedTransactions {
  data: TokenTransaction[];
  total: number;
  limit: number;
  offset: number;
}

export type RewardType = (typeof RewardType)[keyof typeof RewardType];

export const RewardType = {
  USAGE: "USAGE",
  REFERRAL: "REFERRAL",
  CONTRIBUTION: "CONTRIBUTION",
  BONUS: "BONUS",
} as const;

export interface Reward {
  id: string;
  type: RewardType;
  amount: number;
  createdAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  name: string;
  totalRewards: number;
  requestCount: number;
}

export type StakeStatus = (typeof StakeStatus)[keyof typeof StakeStatus];

export const StakeStatus = {
  ACTIVE: "ACTIVE",
  COMPLETED: "COMPLETED",
  WITHDRAWN: "WITHDRAWN",
} as const;

export interface Stake {
  id: string;
  amount: number;
  status: StakeStatus;
  startDate: string;
  endDate?: string | null;
}

export interface CreateStakeBody {
  amount: number;
}

export type ProposalStatus =
  (typeof ProposalStatus)[keyof typeof ProposalStatus];

export const ProposalStatus = {
  ACTIVE: "ACTIVE",
  PASSED: "PASSED",
  REJECTED: "REJECTED",
} as const;

export type VoteChoice = (typeof VoteChoice)[keyof typeof VoteChoice];

export const VoteChoice = {
  YES: "YES",
  NO: "NO",
  ABSTAIN: "ABSTAIN",
} as const;

export interface Proposal {
  id: string;
  title: string;
  description: string;
  status: ProposalStatus;
  endDate: string;
  createdAt: string;
  yesVotes: number;
  noVotes: number;
  abstainVotes: number;
  userVote?: VoteChoice | null;
}

export interface CastVoteBody {
  choice: VoteChoice;
}

export interface Vote {
  id: string;
  proposalId: string;
  choice: VoteChoice;
  weight: number;
  createdAt: string;
}

export interface DashboardSummary {
  requestsToday: number;
  tokenBalance: number;
  totalRewardsEarned: number;
  activeModels: number;
  totalRequests: number;
  tokensSpentToday: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  AI_REQUEST: "AI_REQUEST",
  REWARD_EARNED: "REWARD_EARNED",
  TOKEN_TRANSACTION: "TOKEN_TRANSACTION",
  STAKE: "STAKE",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  description: string;
  amount?: number | null;
  createdAt: string;
}

export interface ToolUsageSummary {
  toolId: string;
  toolName: string;
  toolSlug: string;
  category: AICategory;
  requestCount: number;
  tokensSpent: number;
}

export type ListAIToolsParams = {
  category?: AICategory;
  providerId?: string;
  search?: string;
};

export type ListMyUsageParams = {
  toolId?: string;
  status?: UsageStatus;
  limit?: number;
  offset?: number;
};

export type ListTokenTransactionsParams = {
  type?: TransactionType;
  limit?: number;
  offset?: number;
};

export type ListProposalsParams = {
  status?: ProposalStatus;
};

export type GetDashboardActivityParams = {
  limit?: number;
};