# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI1Net API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /usage:
    get:
      operationId: listMyUsage
      tags: [ai-usage]
      summary: List current user AI usage history
      security:
        - clerkAuth: []
      parameters:
        - name: toolId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/UsageStatus"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Usage history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsage"
  /usage/stats:
    get:
      operationId: getUsageStats
      tags: [ai-usage]
      summary: Get aggregated usage statistics for current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Usage stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageStats"
  /tokens/balance:
    get:
      operationId: getTokenBalance
      tags: [tokens]
      summary: Get current user token balance
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Token balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenBalance"
  /tokens/transactions:
    get:
      operationId: listTokenTransactions
      tags: [tokens]
      summary: List token transaction history
      security:
        - clerkAuth: []
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/TransactionType"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTransactions"
...
  /stakes:
    get:
      operationId: listMyStakes
      tags: [stakes]
      summary: List current user stakes
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Stakes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stake"
    post:
      operationId: createStake
      tags: [stakes]
      summary: Stake tokens
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStakeBody"
      responses:
        "201":
          description: Stake created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stake"
  /governance/proposals:
    get:
      operationId: listProposals
      tags: [governance]
      summary: List governance proposals
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProposalStatus"
      responses:
        "200":
          description: Proposals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
  /governance/proposals/{id}/vote:
    post:
      operationId: castVote
      tags: [governance]
      summary: Cast a vote on a proposal
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CastVoteBody"
      responses:
        "201":
          description: Vote cast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vote"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats for current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent activity feed for current user
      security:
        - clerkAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/tools-usage:
    get:
      operationId: getToolsUsageSummary
      tags: [dashboard]
      summary: Get per-tool usage counts for current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Tools usage summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolUsageSummary"
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    AICategory:
      type: string
      enum: [TEXT, IMAGE, VIDEO, CODE, VOICE, MULTIMODAL]
...
    UsageStatus:
      type: string
      enum: [SUCCESS, FAILED, PENDING]
    AIUsage:
      type: object
      properties:
        id:
          type: string
        toolId:
          type: string
        toolName:
          type: string
        toolSlug:
          type: string
        category:
          $ref: "#/components/schemas/AICategory"
        input:
          type: string
        output:
          type: string
          nullable: true
        tokensUsed:
          type: number
        status:
          $ref: "#/components/schemas/UsageStatus"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - toolId
        - toolName
        - toolSlug
        - category
        - input
        - tokensUsed
        - status
        - createdAt
    PaginatedUsage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AIUsage"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - data
        - total
        - limit
        - offset
    UsageStats:
      type: object
      properties:
        totalRequests:
          type: integer
        totalTokensSpent:
          type: number
        requestsToday:
          type: integer
        tokensSpentToday:
          type: number
        successRate:
          type: number
        topCategories:
          type: array
          items:
            type: object
            properties:
              category:
                $ref: "#/components/schemas/AICategory"
              count:
                type: integer
            required:
              - category
              - count
      required:
        - totalRequests
        - totalTokensSpent
        - requestsToday
        - tokensSpentToday
        - successRate
        - topCategories
    TransactionType:
      type: string
      enum: [SPEND, EARN, STAKE, UNSTAKE, REWARD]
    TokenBalance:
      type: object
      properties:
        balance:
          type: number
        stakedAmount:
          type: number
        availableBalance:
          type: number
      required:
        - balance
        - stakedAmount
        - availableBalance
    TokenTransaction:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          type: number
        metadata:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - amount
        - createdAt
    PaginatedTransactions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TokenTransaction"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - data
        - total
        - limit
        - offset
...
    StakeStatus:
      type: string
      enum: [ACTIVE, COMPLETED, WITHDRAWN]
    Stake:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        status:
          $ref: "#/components/schemas/StakeStatus"
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - amount
        - status
        - startDate
    CreateStakeBody:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount
    ProposalStatus:
      type: string
      enum: [ACTIVE, PASSED, REJECTED]
    VoteChoice:
      type: string
      enum: [YES, NO, ABSTAIN]
    Proposal:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/ProposalStatus"
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        yesVotes:
          type: number
        noVotes:
          type: number
        abstainVotes:
          type: number
        userVote:
          $ref: "#/components/schemas/VoteChoice"
          nullable: true
      required:
        - id
        - title
        - description
        - status
        - endDate
        - createdAt
        - yesVotes
        - noVotes
        - abstainVotes
    CastVoteBody:
      type: object
      properties:
        choice:
          $ref: "#/components/schemas/VoteChoice"
      required:
        - choice
    Vote:
      type: object
      properties:
        id:
          type: string
        proposalId:
          type: string
        choice:
          $ref: "#/components/schemas/VoteChoice"
        weight:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - proposalId
        - choice
        - weight
        - createdAt
    DashboardSummary:
      type: object
      properties:
        requestsToday:
          type: integer
        tokenBalance:
          type: number
        totalRewardsEarned:
          type: number
        activeModels:
          type: integer
        totalRequests:
          type: integer
        tokensSpentToday:
          type: number
      required:
        - requestsToday
        - tokenBalance
        - totalRewardsEarned
        - activeModels
        - totalRequests
        - tokensSpentToday
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [AI_REQUEST, REWARD_EARNED, TOKEN_TRANSACTION, STAKE]
        title:
          type: string
        description:
          type: string
        amount:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - description
        - createdAt
    ToolUsageSummary:
      type: object
      properties:
        toolId:
          type: string
        toolName:
          type: string
        toolSlug:
          type: string
        category:
          $ref: "#/components/schemas/AICategory"
        requestCount:
          type: integer
        tokensSpent:
          type: number
      required:
        - toolId
        - toolName
        - toolSlug
        - category
        - requestCount
        - tokensSpent