# lib/api-zod/src/generated/types/aICategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI1Net API specification
 * OpenAPI spec version: 0.1.0
 */

export type AICategory = (typeof AICategory)[keyof typeof AICategory];

export const AICategory = {
  TEXT: "TEXT",
  IMAGE: "IMAGE",
  VIDEO: "VIDEO",
  CODE: "CODE",
  VOICE: "VOICE",
  MULTIMODAL: "MULTIMODAL",
} as const;